requirements.txt
Django>=1.11.20,<2
semver>=2.8.1,<3
django-nose>=1.4.6,<2
django-webtest>=1.9.4
webtest>=2.0.33
mock>=1.3.0,<2
django-allauth>=0.39.1,<=0.40
django-datatables-view==1.17.0
psycopg2>=2.7,<2.8 --no-binary psycopg2
django-bootstrap3>=8.2.1,<9
requests>=2.21.0,<3
pyvirtualdisplay>=0.2.1,<2
selenium==2.49.0
behave>=1.2.6,<2
selenic>=0.11.0,<1
pebble>=4.3.10,<5
django-redis>=4.10.0,<5
django-cache-management>=3,<4
celery[redis]>=4.2.0,<4.3
wedutil==0.21.3
djangorestframework>=3.9.2,<4
python-slugify>=3,<4
django-cms>=3.6.0,<3.7
# Optional, recommended packages for django-cms
Pillow>=5.4.1
django-filer==1.4.4
cmsplugin-filer==1.1.3
djangocms-text-ckeditor>=3.7.0
cmsplugin-iframe>=0.1.5
lxml>=4,<5
cssselect>=1.0.3,<2
tatsu>=4.3.0,<5
django-pipeline>=1.6.14,<1.7
django-debug-toolbar==1.4
pyexistdb>=0.21.1
eulxml>=1.1.3,<2
pyee>=5,<6
nose-progressive>=1.5.2
glerbl==0.1.1
